resolved #1: extremely bad formal-test-dataset results
It was a stupid bug in datasets.make_formal_test_data() in the call to fit_polynomial().  I was assigning the verbose flag (True) to the reset_const_coeff arg, the result of which is the linear coefficient wasn't being subtracted from the delta mags (so not rectifying them to zero).  DOH!