updated make_formal_test_dataset to calculate labels for sini and cosi
These are calculated and added to the labels data (if not already in the labels block of the config). Whether they are stored is still under the control of the serialize function in deb_example. This brings this function in line with the other trainset & dataset codes which all make sini and cosi available for use as labels.

Related to #5