hyperparam search is now usable (resolves #12)
Reworked the structure of the trials param space so that indididual trial dicts print out clearly and we know what args are being given to each function/class. This is complemented with the new get_trial_value() which can be used to extract items from a trial dict including executing function/class calls in the new grammar.

Some temp model helper funcs which preview a probably change to how the modelling build_ functions work.

Now saves log, best model and best params to the drop directory.