made training/testing independent of ext_features/labels used
This is a scarily big change but it had to be done in one hit.

Added support to the model for generic metadata (in Ouput layer) so that details of features & labels are held within the model, written by the training code and read by the Estimator.  Testing code now gets features and labels information from the Estimator rather than directly from hard coded values in deb_example.

The end result; so long as Estimator can get the metadata from the model, code that interacts with the estimator can find out about feaures, labels, mags_bins and mags_wrap_phase at runtime from the Estimator. This means we can easily switch between models with different feature requirements or predicted parameters.