replaced datasets.inspect_dataset() with deb_example.iterate_dataset()
The inspect_dataset() function was misplaced, and should really have been in the deb_example module as it has too much knowledge of the low level dataset/deb_example structure. It's now been replaced with iterate_dataset() in deb_example which does a similar job except it yields a tuple of ndarrays rather than a tuple of dicts for each row.  It's easy for the calling code to generate dicts, if that's what is needed, with dict(zip(names, array)).