Support for the interactive tester running task 3, 8 or 9 (resolves #76)
jktebop.py code and templates updated so that the same template is used for tasks 3, 8 and 9. Also support for new $task and $simulations tokens in the shared template.

model_testing updated to support this with base_jktebop_task3_params() becoming base_jktebop_fit_params() with args for tasks and simulations.

interactive_tester updated with variables for fit_task and fit_simulation, hooks for accessing that par file written by fitting and conditional code around the fitted model & residuals plot.