dataset histograms now support any csv file names (relates to #79)
Previously the plot_trainset_histogram() function was hard coded to look for trainset*.csv files. Its replacement now takes an iterable of files, so it's up to the caller to glob the appropriate files (which can now be called whatever is wanted).

Took the opportunity to further revise this to return a fig as the other plot functions in this module do.