refactored check for usable instances (relates to #79)
The check for usable instances is now handled as a callback from make_dataset_file() to a function in the make_* modules. Previously this effectively the other way round; a function in datasets callled by the modules.

This change will allow the criteria to be specific to the dataset and for the check to be used within the logic of the controlling make_dataset_file() function.