iterate_dataset() now uses create_map_func()
Now only a single set of code to maintain for handling mapping deb_example as part of a dataset pipeline. This will make it easier to keep things consistent between training, which uses a dataset directly, and the testing code which uses iterate_dataset() to yield a row at a time.