fit_target() supports retry on "good fit not found..." msg (relates to #85)
fit_target() now has a retry_on_failure_to_converge arguments (defaults to True).  When set, the function will capture any "good fit was not found" warnings raised from jktebop module and trigger a second fit taking the fitted output of the first fit as the inputs to the second. The values returned from fit_target() will be from the second fit if it was needed and completed without its own "good fit was not found" warning otherwise it will be from the first fit.