fixed repeatability of testing predictions with MC Dropout (resolves #89)
The seed arg of Estimator.predict() was not having the desired effect. I've found a way of forcing a seed onto the Dropout layers of the estimator's model, but it's a bit of a hack and I'd rather not rely for general use. So, I've removed the seed arg and instead added code specifically to model_testing to find the Dropout layers and set the seeds, prior to making predictions. This is now giving me repeatable results.