we now write error bars to predictions & fit tables where useful  (resolves #88)
predictions_vs_labels_to_table() now supports an error_bars arg (default False). When True the func will write out 2 lines for each row of values, the second headed +/- will be the stddev value any UFloats in the row (otherwise zero).

We set this new arg to True for full (model_tester) testing for prediction text files (if preds are MC) and for fitted param text files. We leave what's written to stdout/log as before, without errorbars.

Also updated quick_fit tool to use this when comparing the fit to the predictions.