better handling of scaled preds/labels
The evaluate_model_against_dataset() func has been fixed to scale inc if it is calculated and scale arg is True. Now also outputs a list of the scales applied so that it's clearer what the subsequent tables and plots relate to.

Removed the reverse_scaling args from predictions_vs_labels_to_table() and calculate_prediction_errors() as they're never used and their presence was confusing.