resolved #91 in interactive tester where labels for disabled targs not read
cleaned up some other code so that the tables for predicted and fitted values vs labels more closely align with those reported by model_testing. Also fixed append_calculated_inc_predictions() as it was returning None if the "inc" column was already present which was a problem if you ran the cell which called it >1 time.