reworked fit_formal_test_dataset() for consistency
Now works in a way which is consistent with the model_interactive_tester.ipynb; each target system is predicted individually and the predictions and fitting is made against directly downloaded/opened fits files. The tfrecord files are not used.

This means that so long as the same seeds are applied the results from fit_formal_test_dataset() should be consistent with this from the model_interactive_tester.ipynb