enable pred-v-label plot for "very low k preds"
Useful as this shows that the same instances are responsible for the horizontal band of poor k preds, the vertical band in rA+rB preds around 0.22, and the horizontal bands of J and bP near 0 and the majority of the scatter on esinw. These are all, bar 1, shallow eclipse instances.