revised training set
Moved to k and J being generated from the ratio of pairs of uniform distributions.

Moved inc to be uniform over cos(inc)

Increased sigma on ecc to 0.5; uniform distribution on this "doesn't work" as we end up with a bias towards extremely eccentric systems. With half-normal(0, 0.5) we end up with close to uniform distribution over [0, 1).