revised training dataset + changes to training to reflect this
Reverted inc to simple uniform dist - simpler. We worry about unifrom over cosi for the mist-tess dataset which needs to reflect a real population.

Also changed what's written so we have a full train/val/test split.  Previously it was 80/20/00 and now it's 80/10/10. This will make it easier to set up the model search. For the full training we can have the validation set as val+test to have the same split as before. Updated the training and search modules to reflect this.