RE calc now uses |(lab-pred)/std(lab)| for ecosw & esinw (resolves #122)
The fiddly aspect of this was to ensure that std(lab) was taken over the entire dataset, even when reporting on a masked subset. Had to calculated the std(lab) values outside the predictions_vs_labels_to_table() (new std_dev func required to handle UFloats) and pass them in as an arg.